+++ src/expect/value_type.mjs
@@ -24,6 +24,14 @@ export function isUndefined(given: mixed) {
   return Object.is(given, undefined)
 }
 
+export function isFunction(given: mixed) {
+  return typeof given === 'function'
+}
+
+export function isPromise(given: mixed) {
+  return given instanceof Promise
+}
+
 export function isArray(given: mixed) {
   return Array.isArray(given)
 }
@@ -41,4 +49,17 @@ export function isMap(given: mixed) {
 
 export function isWeakMap(given: mixed) {
   return given instanceof WeakMap
+}
+
+export function isPlainObject(given: mixed) {
+  return (
+    typeof given === 'object' &&
+    !isNull(given) &&
+    !isArray(given) &&
+    !isSet(given) &&
+    !isWeakSet(given) &&
+    !isMap(given) &&
+    !isWeakMap(given) &&
+    !isPromise(given)
+  )
 }